HT_21/requirements.txt
Django==5.0
django-autoslug==1.9.9
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
requests==2.31.0